# generated/src/main/java/com/aimyapp/api/DelayApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.23.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.aimyapp.api;

import com.aimyapp.model.DelayNotifyRequest;
import com.aimyapp.model.DelayNotifyResponse;
import com.aimyapp.model.ErrorResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-06-28T15:16:53.729415900+09:00[Asia/Tokyo]", comments = "Generator version: 7.23.0")
@Validated
@Tag(name = "Delay", description = "the Delay API")
public interface DelayApi {

    String PATH_NOTIFY_DELAY = "/delay/notify";
    /**
     * POST /delay/notify : 遅延情報通知
     *
     * @param delayNotifyRequest  (required)
     * @return 通知成功 (status code 200)
     *         or リクエストが不正です (status code 400)
     *         or 予期せぬエラーが発生しました (status code 500)
     */
    @Operation(
        operationId = "notifyDelay",
        summary = "遅延情報通知",
        tags = { "Delay" },
        responses = {
            @ApiResponse(responseCode = "200", description = "通知成功", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DelayNotifyResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "リクエストが不正です", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "予期せぬエラーが発生しました", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = DelayApi.PATH_NOTIFY_DELAY,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    ResponseEntity<DelayNotifyResponse> notifyDelay(
        @Parameter(name = "DelayNotifyRequest", description = "", required = true) @Valid @RequestBody DelayNotifyRequest delayNotifyRequest
    );

}