# openapi.yaml
openapi: 3.0.3

info:
  title: Sample API
  version: 1.0.0

paths:
  /hello:
    get:
      operationId: hello
      tags:
        - Hello
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HelloResponse"

  /delay/notify:
    post:
      tags:
        - Delay
      operationId: notifyDelay
      summary: 遅延情報通知

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DelayNotifyRequest'

      responses:
        '200':
          description: 通知成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DelayNotifyResponse'

        '400':
          description: リクエストが不正です
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: 予期せぬエラーが発生しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


components:
  schemas:

    HelloResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string


    DelayNotifyRequest:
      type: object
      required:
        - from
        - to
      properties:
        from:
          type: string
          description: 出発駅
          example: "東京"

        to:
          type: string
          description: 到着駅
          example: "新宿"


    DelayNotifyResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          example: "SUCCESS"


    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: エラーコード
          example: "INVALID_REQUEST"

        message:
          type: string
          description: エラーメッセージ
          example: "リクエスト内容が不正です"

        timestamp:
          type: string
          format: date-time
          example: "2026-06-28T12:00:00+09:00"